+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Version 0.6.0:
 - Added support to remove an event from a repeating series (asked in the confirmation message).
+- Holidays (Christmas Eve, Christmas Day, Boxing Day, New Year's Eve and New Year's Day) are now the default holidays.
 
 ### Version 0.5.2:
 - The text "All Day Event" is now "All-Day" (renamed all the settings to match).